extend BindAddress option to select the type of IPv6 address
When using IPv6 autoconfiguration, the IP address is deducted using the unique
hardware address of the network card and the announced network prefix. Since
this might lead to privacy issues, most operating systems generate
pseudo-random addresses that are rotated in regular intervals.

This can be a problem for long-running connections if a address is invalidated
while still in use - the connection "hangs". Even though expired addresses are
usually retained for a long timeframe to prevent this, accidental dis- and
reconnection (e.g. when using a wireless network) flushes the list of
previously used addresses.

By setting appropiate socket options, the kernel can be instructed to use the
public (and static) source address for the outgoing connection. This change
implements this functionality for SSH, adding special behaviour to the option
"BindAddress" (-b) that can be set to "%pub(lic)", "%t(e)mp", indicating the
preference for the type of address to be used. The prefixed percentage sign is
used to distinguish the arguments from ordinary hostnames (just like ping6 does
with appended interface names on link local addresses).