add environment variable SSH_CUSTOM_CMD for ProxyJump
This change makes the ProxyJump functionality read the environment variable
SSH_CUSTOM_CMD to construct the proxy command line; this makes it possible to
specify the path to a custom ssh binary as well as add extra flags. Even if the
variable is unspecified or empty, the ssh binary used to establish the
conection is not necessarily the one found first in $PATH; this patch uses the
file /proc/self/exe available on many platforms to identify the binary of the
already running process to launch a ProxyJump forwarding.

If these two approaches fail, it falls back to the existing behaviour of just
assuming "ssh" as command.